feat: enhance vnstat agent list UI with integrated controls
- Merge agent list design with details header for better visual hierarchy
- Add ServerStackIcon and all agent info (URL, retention, live rates) in list items
- Move Start/Stop button directly to agent list for easier access
- Simplify VnstatAgentDetails to remove duplicate header info
- Add instant star toggle feedback using local state
- Enable agent deselection by clicking selected agent
- Improve dark mode styling consistency
- Fix table border colors in VnstatUsageTable
